Parallelized step --merge
Added support for executing in parallel the merge step
Added the method createDirectoryHierarchy to StackStitcher to create
only the directories of the output (stitched) volume.
Added the option —makedirs to execute the above method.
Added the method mdataGenerator to StackStitcher to generate the
metadata file of the output (stitched) volume.
Added the option —metadata to enable the execution of the above method
Added parameter par_mode to the method  mergeTilesVaa3DRaw to disable
side-effects (to be used when more instances of the method are executed
in parallel)